Added: export to file function.
Through export command, allow the user to export the results to a file.
It is useful when analyzing multiple samples (enough to making scroll up
impossible) or just to save results.
Path completion is provided for input and then ishell completion is
reset.
Asks for a confirmation is file already exists.